+++ Sources/SwiftUICharts/BarChart/Views/GroupedBarChart.swift
@@ -74,7 +74,7 @@ public struct GroupedBarChart<ChartData>: View where ChartData: GroupedBarChartD
             } else { CustomNoDataView(chartData: chartData) }
         }
         .if(chartData.minValue.isLess(than: 0)) {
-            $0.scaleEffect(y: chartData.maxValue/(chartData.maxValue - chartData.minValue), anchor: .top)
+            $0.scaleEffect(y: CGFloat(chartData.maxValue/(chartData.maxValue - chartData.minValue)), anchor: .top)
         }
     }
 }
